+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Dart Package Versioning](https://dart.dev/tools/pub
 
 ## [Unreleased]
 
+### Changed
+
+- drop the dependency in the "cpf_cnpj_validator" package —
+  [104](https://github.com/dartoos-dev/formdator/issues/104).
+
 ## [0.8.1] - 2021-07-26
 
 ### Added